+++ django/applications/catmaid/urls.py
@@ -413,6 +413,7 @@ urlpatterns += [
     url(r'^(?P<project_id>\d+)/similarity/queries/similarity$', similarity.compare_skeletons),
     url(r'^(?P<project_id>\d+)/similarity/queries/(?P<similarity_id>\d+)/$', similarity.SimilarityDetail.as_view()),
     url(r'^(?P<project_id>\d+)/similarity/queries/(?P<similarity_id>\d+)/recompute$', similarity.recompute_similarity),
+    url(r'^(?P<project_id>\d+)/similarity/queries/(?P<similarity_id>\d+)/storage$', similarity.SimilarityStorageDetail.as_view()),
     url(r'^(?P<project_id>\d+)/similarity/test-setup$', similarity.test_setup),
 ]